requirements.txt
colorama==0.4.6
discord.py==2.3.2
